Updated #glow(boolean) to disenchant
Previously, this would just leave the item in its original state. Therefore, if the item was enchanted before the glow(false) was redundant and kept the item enchanted. Now it will respect the boolean and disenchant the item too. 

Note: 
As far as I know it's impossible to solely remove the glow factor without removing enchantments as they are linked.
Thus add an enchant = glow, remove an enchant = no glow. There's no way to remove an enchant and keep the glow to my knowledge.